requirements.txt
pandas==0.24.2
pytest==5.0.0
PyPDF2==1.26.0
python-docx==0.8.10